Update go versions throughout Pelican repo
Updates to a few underyling packages caused `go mod tidy` to pick up
a new version of Go without my explicitly having changed it.

Looks like we hardcode the go version in a half dozen or so different
places, and these are required for lots of tests to run correctly.
Updating things across the board.